Simplify primary app texts
To improve UX and onboarding, MPW needs to simplify it's text, reduce it, and just generally make it a little dumber.

This commit massively reduces the Dashboard text, fixes some i18n tags/text, removes an outdated Staking message and replaces it with a more useful explanation of the Staking feature.